docs(problem5): 각 클래스와 메소드의 역할을 설명하는 Javadoc 추가
